Update comment addition process and display
Modified the 'addComment' method in the CommentController to directly fetch an updated list of the comments after adding a new comment rather than returning the added comment details. In the frontend, added a form submission functionality to 'ShowPost' component and 'usePost' hook to facilitate users to post a comment through the form. This change enhances the user experience as the comments list is refreshed automatically after adding a new comment.